마을 디자인 추가

+++ RhythmiCal/src/main/webapp/WEB-INF/views/town.jsp
@@ -11,7 +11,7 @@
 <script>
 var game = new Phaser.Game(1600, 900, Phaser.AUTO, 'phaser-example', { preload: preload, create: create, update: update});
 
-var text1;
+var text1,text2;
 var cursors;
 var point;
 var image;
@@ -25,9 +25,11 @@ var bgd;
 
 var key; // 키보드 버튼
 var depth; // 작업소 깊이
-var squareX, squareY; // 스퀘어의 위치(좌표)를 나타내는 변수
+var square,squareX, squareY; // 스퀘어의 위치(좌표)를 나타내는 변수
+var buttonX=1000, buttonY=450; // 모션, 효과, 레인 스프라이트의 위치(좌표)를 나타내는 변수
 var buttonFocus; // 현재 버튼이 몇 번에 있는가
 var motion; // 모션을 담는 배열
+var m_back; // 서브메뉴 배경
 var turn1, turn2, turn3, turn4, turn5, turn6; // 모션(1~3) 및 레인(4~6)을 교체할 카운터
 var motion1, motion2, motion3, effect1, effect2, effect3, lane1, lane2, lane3; // 9개의 스프라이트 버튼
 var singleLane, doubleLane; // 싱글레인 및 더블레인 배열
@@ -46,12 +48,17 @@ function preload() {
 	
 	// 진주 이미지	
 	game.load.image('menuwin','resources/Images/town/townImg/vmenu.png' ); //마을 메뉴 이미지 로드
-	game.load.image('click','resources/Images/town/townImg/click.jpg' ); //첫번째 메뉴 이미지 로드(임의)
+	game.load.image('myroom','resources/Images/town/townImg/myroom.png' ) // 내방 이미지
 	game.load.image('click2','resources/Images/town/townImg/click2.png' ); //두번째 메뉴 이미지 로드(임의)
 	game.load.image('hand','resources/Images/town/townImg/hand.png' ); //스마트폰 들고있는 이미지 로드
 	game.load.image('finish','resources/Images/town/townImg/black.png' ); //종료시 fade out될 검정 배경 이미지 로드
 	game.load.image('back','resources/Images/town/townImg/v_back.png' ); //마을 배경 이미지
 	game.load.image('select','resources/Images/town/townImg/select.png' );//선택 흰 테두리
+	game.load.image('menu_back','resources/Images/town/townImg/menu_back.png' );//Enter 눌렀을 때 서브메뉴 배경
+	game.load.image('pub','resources/Images/town/townImg/pub.PNG' ); //용병소 이미지
+	game.load.image('worksplace', 'resources/Images/town/townImg/office.png'); //작업소 이미지
+	game.load.image('exit','resources/Images/town/townImg/exit.png' ); //내방에서의 종료 버튼 이미지
+	game.load.image('e_select','resources/Images/town/townImg/exit_line.png' ); //종료 버튼 감싸고 있는 선택 이미지
 	
 	// 네모 테두리 로드
 	game.load.spritesheet('square', 'resources/Images/town/produceRoom/square.png', 95, 95);
@@ -178,11 +185,11 @@ function moveMenu() {
 		} break;
 	case 'enter':
 		isEntered = true;
-	   	point.kill();
-	   	point = null;
+		if (point != null) {
+			point.kill();
+		   	point = null;			
+		}
 	   	
-	   	//취소키 (1번이라고 가정)
-	   	key1 = game.input.keyboard.addKey(Phaser.Keyboard.ONE);
 	   	
 	   	// 화살표가 멈춰있는 위치에서 엔터를 눌렀을 때 분기 처리.
 		switch (y) {
@@ -201,15 +208,7 @@ function moveMenu() {
 				break;
 			case 540: alert('겜종료');
 				isnull();
-				// 게임 종료. 검정 화면 준비.
-				var sprite = game.add.sprite(0, 0, 'finish');
-				// 원래 사이즈 보다 확대 하고 alpha로 투명도 조절.
-				sprite.scale.set(5);
-			    sprite.anchor.setTo(0.5, 0.5);
-			    sprite.alpha = 0;
-			    //화면에서 검정화면으로 조정.
-				game.add.tween(sprite).to( { alpha: 1 }, 2000, Phaser.Easing.Linear.None, true, 0, 0, false);
-			    //로컬로 이동해서 로그아웃 등을 해주는 작업 필요.
+				myroom();
 			    logoutMember();
 				break;
 		}
@@ -233,9 +232,8 @@ function logoutMember() {
  * createStudio(): 작업소 화면을 만드는 메소드 
  */
 function createStudio() {
-	image = game.add.image(600, 80, 'click');
-	key1.onDown.add(cancel, this); 
-	
+	image = game.add.image(770, 120, 'worksplace');
+
 	// 버튼 포커스를 1로 초기화
 	buttonFocus = 1;
 	// 순서를 0으로 초기화
@@ -259,7 +257,7 @@ function createStudio() {
 	motion = [Point, Up, Down, Left, Right];
 	
 	// 스퀘어 생성
-	squareX = 1000; squareY = 200;
+	squareX = 1000; squareY = 450;
 	square = game.add.sprite(squareX, squareY, 'square');
 	
 	// AJAX를 통해 DB(table save)로부터 모션 리스트를 읽음
@@ -270,36 +268,36 @@ function createStudio() {
 		success: function(json) {
 			var motionList = $.parseJSON(json);
 			// 첫 번째 모션
-			motion1 = game.add.sprite(1000, 200, motionList[0].name);
-			effect1 = game.add.sprite(1000, 300, motionList[0].effect);
-			lane1 = game.add.sprite(1000, 400, motionList[0].lane);
+			motion1 = game.add.sprite(buttonX, buttonY, motionList[0].name);
+			effect1 = game.add.sprite(buttonX, buttonY+100, motionList[0].effect);
+			lane1 = game.add.sprite(buttonX, button+200, motionList[0].lane);
 			
 			// 두 번째 모션
-			motion2 = game.add.sprite(1100, 200, motionList[1].lane);
-			effect2 = game.add.sprite(1100, 300, motionList[1].lane);
-			lane2 = game.add.sprite(1100, 400, motionList[1].lane);
+			motion2 = game.add.sprite(buttonX+100, buttonY, motionList[1].lane);
+			effect2 = game.add.sprite(buttonX+100, buttonY+100, motionList[1].lane);
+			lane2 = game.add.sprite(buttonX+100, buttonY+200, motionList[1].lane);
 			
 			// 세 번째 모션
-			motion3 = game.add.sprite(1200, 200, motionList[2].lane);
-			effect3 = game.add.sprite(1200, 300, motionList[2].lane);
-			lane3 = game.add.sprite(1200, 400, motionList[2].lane);
+			motion3 = game.add.sprite(buttonX+200, buttonY, motionList[2].lane);
+			effect3 = game.add.sprite(buttonX+200, buttonY+100, motionList[2].lane);
+			lane3 = game.add.sprite(buttonX+200, buttonY+200, motionList[2].lane);
 	   },
 	   // 실패하면 기본값을 표시
 	   error: function() {
 			// 첫 번째 모션
-			motion1 = game.add.sprite(1000, 200, 'point');
-			effect1 = game.add.sprite(1000, 300, 'fire');
-			lane1 = game.add.sprite(1000, 400, 'A');
+			motion1 = game.add.sprite(buttonX, buttonY, 'point');
+			effect1 = game.add.sprite(buttonX, buttonY+100, 'fire');
+			lane1 = game.add.sprite(buttonX, buttonY+200, 'A');
 			
 			// 두 번째 모션
-			motion2 = game.add.sprite(1100, 200, 'up');
-			effect2 = game.add.sprite(1100, 300, 'water');
-			lane2 = game.add.sprite(1100, 400, 'A');
+			motion2 = game.add.sprite(buttonX+100, buttonY, 'up');
+			effect2 = game.add.sprite(buttonX+100, buttonY+100, 'water');
+			lane2 = game.add.sprite(buttonX+100, buttonY+200, 'A');
 			
 			// 세 번째 모션
-			motion3 = game.add.sprite(1200, 200, 'down');
-			effect3 = game.add.sprite(1200, 300, 'sun');
-			lane3 = game.add.sprite(1200, 400, 'A');
+			motion3 = game.add.sprite(buttonX+200, buttonY, 'down');
+			effect3 = game.add.sprite(buttonX+200, buttonY+100, 'sun');
+			lane3 = game.add.sprite(buttonX+200, buttonY+200, 'A');
 	   }
 	});
 	  
@@ -349,8 +347,12 @@ function moveButtonFocus() {
     	   	 depth = 2; 
     	  	 break;
       case 'esc': 
-    	  	 point = game.add.image(x, y, 'select');
-  		 point.scale.set(1.98);
+  		isEntered = false;
+  		if (point == null) {
+  			point = game.add.image(x,y,'select');	
+  			point.scale.set(1.98);
+  		} 
+  		
   		 depth = 0; 
   		 break;
    }
@@ -369,17 +371,17 @@ function moveContent(buttonFocus) {
 	         if (turn1 == 0) {return;} turn1 = turn1-1;
 	         if (motion[turn1].getName() == motion2.key || motion[turn1].getName() == motion3.key) {return;}
 	         else {
-	        	 	 motion1 = game.add.sprite(1000, 200, motion[turn1].getName()); 
-	             effect1 = game.add.sprite(1000, 300, motion[turn1].getEffect());
-	             lane1 = game.add.sprite(1000, 400, motion[turn1].getLane()[0]);
+	        	 motion1 = game.add.sprite(buttonX, buttonY, motion[turn1].getName()); 
+	             effect1 = game.add.sprite(buttonX, buttonY+100, motion[turn1].getEffect());
+	             lane1 = game.add.sprite(buttonX, buttonY+200, motion[turn1].getLane()[0]);
 	         } break;
 	      case 'right':
 	         if (turn1 >= 4) {turn1 = 3;} turn1 = turn1+1;
 	         if (motion[turn1].getName() == motion2.key || motion[turn1].getName() == motion3.key) {return;}
 		     else {
-		    	 	 motion1 = game.add.sprite(1000, 200, motion[turn1].getName()); 
-		         effect1 = game.add.sprite(1000, 300, motion[turn1].getEffect());
-		         lane1 = game.add.sprite(1000, 400, motion[turn1].getLane()[0]);
+		    	 motion1 = game.add.sprite(buttonX, buttonY, motion[turn1].getName()); 
+		         effect1 = game.add.sprite(buttonX, buttonY+100, motion[turn1].getEffect());
+		         lane1 = game.add.sprite(buttonX, buttonY+200, motion[turn1].getLane()[0]);
 			 } break;
 	      case 'enter': 
 	         //이 모션으로 선택했다는 효과 주기
@@ -394,17 +396,17 @@ function moveContent(buttonFocus) {
 	         if (turn2 == 0) {return;} turn2 = turn2-1;
 	         if (motion[turn2].getName() == motion1.key || motion[turn2].getName() == motion3.key) {return;}
 	         else {
-		        	 motion2 = game.add.sprite(1100, 200, motion[turn2].getName());
-	             effect2 = game.add.sprite(1100, 300, motion[turn2].getEffect());
-	             lane2 = game.add.sprite(1100, 400, motion[turn2].getLane()[0]);
+		         motion2 = game.add.sprite(buttonX+100, buttonY, motion[turn2].getName());
+	             effect2 = game.add.sprite(buttonX+100, buttonY+100, motion[turn2].getEffect());
+	             lane2 = game.add.sprite(buttonX+100, buttonY+200, motion[turn2].getLane()[0]);
 	         } break;
 	      case 'right': 
 	         if (turn2 >= 4) {turn2 = 3;} turn2 = turn2+1;
 	         if (motion[turn2].getName() == motion1.key || motion[turn2].getName() == motion3.key) {return;}
 		     else {
-		        	 motion2 = game.add.sprite(1100, 200, motion[turn2].getName());
-		         effect2 = game.add.sprite(1100, 300, motion[turn2].getEffect());
-		         lane2 = game.add.sprite(1100, 400, motion[turn2].getLane()[0]);
+		         motion2 = game.add.sprite(buttonX+100, buttonY, motion[turn2].getName());
+		         effect2 = game.add.sprite(buttonX+100, buttonY+100, motion[turn2].getEffect());
+		         lane2 = game.add.sprite(buttonX+100, buttonY+200, motion[turn2].getLane()[0]);
 		     } break;
 	      case 'enter':
 	         //이 모션으로 선택
@@ -419,17 +421,17 @@ function moveContent(buttonFocus) {
 	         if (turn3 == 0) {return;} turn3 = turn3-1;
 	         if (motion[turn3].getName() == motion1.key || motion[turn3].getName() == motion2.key) {return;}
 	         else {
-	        	 	 motion3 = game.add.sprite(1200, 200, motion[turn3].getName());
-	             effect3 = game.add.sprite(1200, 300, motion[turn3].getEffect());
-	             lane3 = game.add.sprite(1200, 400, motion[turn3].getLane()[0]);
+	        	 motion3 = game.add.sprite(buttonX+200, buttonY, motion[turn3].getName());
+	             effect3 = game.add.sprite(buttonX+200, buttonY+100, motion[turn3].getEffect());
+	             lane3 = game.add.sprite(buttonX+200, buttonY+200, motion[turn3].getLane()[0]);
 	         } break;
 	      case 'right': 
 	         if (turn3 >= 4) {turn3 = 3;} turn3 = turn3+1;
 	         if (motion[turn3].getName() == motion1.key || motion[turn3].getName() == motion2.key) {return;}
 	         else {
-	        	 	 motion3 = game.add.sprite(1200, 200, motion[turn3].getName());
-	             effect3 = game.add.sprite(1200, 300, motion[turn3].getEffect());
-	             lane3 = game.add.sprite(1200, 400, motion[turn3].getLane()[0]);	
+	        	 motion3 = game.add.sprite(buttonX+200, buttonY, motion[turn3].getName());
+	             effect3 = game.add.sprite(buttonX+200, buttonY+100, motion[turn3].getEffect());
+	             lane3 = game.add.sprite(buttonX+200, buttonY+200, motion[turn3].getLane()[0]);	
 	        	 } break;
 	      case 'enter':
 	         //이 모션으로 선택했다는 효과 주기
@@ -443,10 +445,10 @@ function moveContent(buttonFocus) {
       switch (inputKey) {
 	      case 'left':
 	         if (turn4 <= 0) {turn4 = 0; return;} turn4 = turn4-1;
-	        	 lane1 = game.add.sprite(1000, 400, Motion.getLane()[turn4]); break;
+	        	 lane1 = game.add.sprite(buttonX, buttonY+200, Motion.getLane()[turn4]); break;
 	      case 'right': 
 	         if (turn4 >= 2) {turn4 = 2; return;} turn4 = turn4+1;
-	        	 lane1 = game.add.sprite(1000, 400, Motion.getLane()[turn4]); break;
+	        	 lane1 = game.add.sprite(buttonX, buttonY+200, Motion.getLane()[turn4]); break;
 	      case 'enter': 
 	    	  	 //이 레인으로 선택했다는 효과 주기
 	         break;
@@ -459,10 +461,10 @@ function moveContent(buttonFocus) {
       switch (inputKey) {
 	      case 'left': 
 	         if (turn5 <= 0) {turn5 = 0; return;} turn5 = turn5-1;
-	        	 lane2 = game.add.sprite(1100, 400, Motion.getLane()[turn5]); break;
+	        	 lane2 = game.add.sprite(buttonX+100, buttonY+200, Motion.getLane()[turn5]); break;
 	      case 'right':
 	         if (turn5 >= 2) {turn5 = 2; return;} turn5 = turn5+1;
-	        	 lane2 = game.add.sprite(1100, 400, Motion.getLane()[turn5]); break;
+	        	 lane2 = game.add.sprite(buttonX+100, buttonY+200, Motion.getLane()[turn5]); break;
 	      case 'enter': 
 	    	  	 //이 레인으로 선택했다는 효과 주기
 	         break;
@@ -475,10 +477,10 @@ function moveContent(buttonFocus) {
       switch (inputKey) {
 	      case 'left': 
 	         if (turn6 <= 0) {turn6 = 0; return;} turn6 = turn6-1;
-	        	 lane3 = game.add.sprite(1200, 400, Motion.getLane()[turn6]); break;
+	        	 lane3 = game.add.sprite(buttonX+200, buttonY+200, Motion.getLane()[turn6]); break;
 	      case 'right': 
 	         if (turn6 >= 2) {turn6 = 2; return;} turn6 = turn6+1;
-	        	 lane3 = game.add.sprite(1200, 400, Motion.getLane()[turn6]); break;
+	        	 lane3 = game.add.sprite(buttonX+200, buttonY+200, Motion.getLane()[turn6]); break;
 	      case 'enter':
 	         //이 레인으로 선택했다는 효과 주기
 	         break;
@@ -492,13 +494,15 @@ function moveContent(buttonFocus) {
  * createMercenary(): 용병소 화면을 만드는 메소드
  */
 function createMercenary() {
+	image = game.add.image(770, 120, 'pub');
+		
 	// 난수 발급
 	var rdm = Math.floor(Math.random() * 9999) + 1000;
-	image = game.add.image(600, 80, 'hand');
-	// 난수를 보여줄 텍스트
-	text1 = game.add.text(870, 180, rdm, { font: "40px Arial", fill: "#000000", align: "center" });
-	// 1번(취소 버튼이라고 가정) 을 눌렀을 때
-	key1.onDown.add(cancel, this);
+	image = game.add.image(800, 350, 'hand');
+		
+		// 난수를 보여줄 텍스트
+	text1 = game.add.text(1070, 450, rdm, 
+			{ font: "40px Arial", fill: "#000000", align: "center" });
 	// 스마트 폰에서 입력한 값과 값을 비교해서 맞으면 연결 시켜주는 작업 필요.	
 }
 
@@ -526,21 +530,52 @@ function update() {
    }
 }
 
+function myroom() {
+	image = game.add.image(770, 120, 'myroom');
+	exit = game.add.image(900, 550, 'exit');
+	exit.scale.set(0.8);
+	text2 = game.add.text(940, 520, "게임을 종료합니다", 
+			{ font: "40px Arial", fill: "#FFFFFF", align: "center" });
+	
+	
+ 	if (game.input.keyboard.isDown(Phaser.Keyboard.ENTER)) {
+		
+		var e_select = game.add.sprite(900, 550, 'e_select');
+		e_select.scale.set(0.8);
+		// 게임 종료. 검정 화면 준비.
+		var sprite = game.add.sprite(0, 0, 'finish');
+		// 원래 사이즈 보다 확대 하고 alph로 투명도 조절.
+		sprite.scale.set(5);
+	    sprite.anchor.setTo(0.5, 0.5);
+	    sprite.alpha = 0;
+		
+	    //화면에서 검정화면으로 조정.
+		game.add.tween(sprite).to( { alpha: 1 }, 2000, Phaser.Easing.Linear.None, true, 0, 0, false);
+	} 
+}
+
+
 function isnull() {
 	if (text1 != null) {text1.kill();}
 	if (image != null) {image.kill();}
+	if (m_back== null) {m_back = game.add.image(720,80,'menu_back');}
+	
+	if (motion1 != null) {motion1.kill();}
+	if (motion2 != null) {motion2.kill();}
+	if (motion3 != null) {motion3.kill();}
+		
+	if (effect1 != null) {effect1.kill();}
+	if (effect2 != null) {effect2.kill();}
+	if (effect3 != null) {effect3.kill();}
+		
+	if (lane1 != null) {lane1.kill();}
+	if (lane2 != null) {lane2.kill();}
+	if (lane3 != null) {lane3.kill();}
+	if (square != null) {square.kill();}
+	
 }
 
-function cancel(){
-	isEntered = false;
-	if (point == null) {
-		point = game.add.image(x,y,'select');	
-		point.scale.set(1.98);
-	} 
-	else {
-		return;
-	}
-}
+
 </script>
 </body>
 </html>
